// src/app/interfaces/reserva.interface.ts
  export interface Reserva {
      _id?: string;
      desde: string;
      hacia: string;
      fecha_vuelo: string;
      hora_vuelo: string;
      precio_manana: string;
      nombre: string;
      apellido: string;
      cedula: string;
      nacimiento?: string;
      email: string;
      reserva?: string;
}